requirements.txt
python-dotenv==1.0.0
twikit==0.2.0
aiohttp==3.9.1
asyncio==3.4.3 